+++ src/main/java/com/targomo/client/api/geo/AbstractCoordinate.java
@@ -8,18 +8,16 @@ import javax.persistence.*;
  * Simple abstract class to use for storing coordinates with IDs and travel types.
  */
 @MappedSuperclass
-public abstract class AbstractCoordinate implements Coordinate {
-
-	private String id;
+public abstract class AbstractCoordinate extends AbstractLocation implements Coordinate {
 
 	private double x;
 
 	private double y;
 
-	private AggregationInputParameters aggregationInputParameters;
-
 	// needed for jackson
-	public AbstractCoordinate(){}
+	public AbstractCoordinate(){
+		super();
+	}
 
 	/**
 	 * Generate a Coordinate with an ID along with X and Y values.
@@ -28,26 +26,9 @@ public abstract class AbstractCoordinate implements Coordinate {
 	 * @param y Y value of target
 	 */
 	public AbstractCoordinate(final String id, final double x, final double y, final AggregationInputParameters aggregationInputParameters) {
-		this.id = id;
+		super(id, aggregationInputParameters);
 		this.x = x;
 		this.y = y;
-		this.aggregationInputParameters = aggregationInputParameters;
-	}
-
-	/**
-	 * Get the ID associated with the coordinate.
-	 * @return Coordinate ID
-	 */
-	public String getId() {
-		return id;
-	}
-
-	/**
-	 * Assign an ID to the coordinate
-	 * @param id ID to be assigned
-	 */
-	public void setId(final String id) {
-		this.id = id;
 	}
 
 	/**
@@ -81,18 +62,6 @@ public abstract class AbstractCoordinate implements Coordinate {
 	public void setY(final double y) {
 		this.y = y;
 	}
-
-	/**
-	 * Get the aggregation input parameters of a location
-	 * @return Location Properties
-	 */
-	public AggregationInputParameters getAggregationInputParameters() { return this.aggregationInputParameters; }
-
-	/**
-	 * Assign aggregation input parameters to a location
-	 * @param aggregationInputParameters aggregation input parameters to be assigned
-	 */
-	public void setAggregationInputParameters(final AggregationInputParameters aggregationInputParameters){ this.aggregationInputParameters = aggregationInputParameters; }
 
 	/**
 	 * Returns a JSON String representation of the Coordinate with ID, x and y values.
